+++ c/containers/dvector.h
@@ -3,11 +3,11 @@
 
 typedef struct _dvector_t {
 	double* data;
-	int     size;
-	int     capacity;
+	unsigned long long size;
+	unsigned long long capacity;
 } dvector_t;
 
-dvector_t* dvector_alloc(int initial_capacity);
+dvector_t* dvector_alloc(unsigned long long initial_capacity);
 void dvector_free(dvector_t* pdvector);
 void dvector_append(dvector_t* pdvector, double value);
